A Solid night's work
The individual pages for characters and maps are more or less feature complete.
Obviously there's still a lot to do and a ton more visual elements and data to add,
but what's there isn't a bad representation of the final product.

Good progress has been made on the encounter page. There's a lot more to be done
in terms of client-side scripting to, for example, enable counters to be dragged
around the canvas. There's a lot to be done in terms of sizing images appropriately,
making image (pre)loading less clunky, making counters more generic, and adding
interactive features and polish.